Extract ViewFinder to class
